requirements.txt
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
atlassian-python-api>=3.41.0
streamlit>=1.28.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
langchain>=0.0.350
langchain-openai>=0.0.2
tiktoken>=0.5.1
